+++ src/game_interact.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   game_interact.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: psmolin <psmolin@student.42heilbronn.de    +#+  +:+       +#+        */
+/*   By: nmikuka <nmikuka@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/04 10:25:29 by nmikuka           #+#    #+#             */
-/*   Updated: 2025/11/10 02:30:28 by psmolin          ###   ########.fr       */
+/*   Updated: 2025/11/12 19:53:47 by nmikuka          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -96,7 +96,7 @@ void	interact(t_gs *game)
 {
 	t_player	*player;
 	t_vec2		vec_door;
-	float		door_angle;
+	double		door_angle;
 	int			i;
 
 	player = game->player;
@@ -105,7 +105,7 @@ void	interact(t_gs *game)
 	{
 		vec_door = (t_vec2){game->inter_walls[i].sprite.pos.x - player->pos.x,
 			game->inter_walls[i].sprite.pos.y - player->pos.y};
-		door_angle = fabsf(ft_angle_between_vec2(vec_door,
+		door_angle = fabs(ft_angle_between_vec2(vec_door,
 					(t_vec2){player->lookdir.x, player->lookdir.y}));
 		if (ft_vec2_length(vec_door) <= INTERACT_DIST
 			&& door_angle < INTERACT_ANGLE)
